+++ crawler/src/lib.rs
@@ -1,5 +1,6 @@
 mod cache;
 mod core;
+mod http;
 mod instances;
 mod runtime;
 pub mod sdk;